genai-project-insights/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.2.0
anthropic>=0.20.0
openai>=1.14.0
google-generativeai>=0.4.0
gitpython>=3.1.40
watchdog>=4.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
websockets>=12.0
